requirements.txt
Django==2.1.1
django-environ==0.4.5
django-money==0.14.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==5.2.0
psycopg2-binary==2.7.5
py-moneyed==0.7.0
PyJWT==1.6.4
pytz==2018.5
